Day 002 of #100DaysOfCode I learned how to write and read file from Disk, how to randomize something and make test file in Go.
I publishd my first Go Library Package called "Random Generator".

When we want a function to return something, we need to tell that function what kind of data type that it will return
eg:
```
func ReturnString() string {
	// Code here
	return "Hello World"
}

// and if the function is going to return more than 1 value, then it should be like this
func MoreThanOneValue() (string, int) {
	return "Hello World", 69
}
```

Naming convention is a lot different from other language, so yeah, it's kinda feel weird sometimes.
And if a variable that should be declared but you dont want to use it, u have to write it with "_"
eg:
```
for _, card := range cards {
	// code here
}
```