So these two days I've learned about Receiver, structs, and pointer
Receiver and structs are okay, but pointer? It's kinda okay.

Today (Day of 004 of #100) i just learned different between value types and reference types
Value Types:
-	Int
- String
- Bool
- Float
- Structs
meanwhile, Reference Types are:
- Slices
- Maps
- Channels
- Pointers, and
- Function

The biggest difference between these two types is how to update is (As far as I know)
For the value types, we need to use pointer to update it, but for reference types we don't need pointer.
CMIIW,
If we create a reference types' variable, we create two things into our memory
1. It variable infos'
2. It variable value
The variable info is pointing to it value in another memory address.
And when we tryning to update it, It'll create a copy of it variable info to new memory address that pointin to the value
A is the variable info
B is variable value
C is the copy of the variable that created by Go, When we use this memory address, it also use "B" as it value (Pointing at "B")

Well, I'm kinda confused right know, really. I need to ask somebody