Maps
Map is the closest data types in GO that I think is similiar to Array in javascript.
But, in map the keys and values are should consist of the same data types.

I am not saying, Key and Value are the same data types, but
if first key is string then all other key are supposed to be string too, so does with the values

Changes to be committed:
-	modified:   Map/main.go
-	renamed:    sandbox/main.go -> Sandbox/main.go