Interfaces
eg

```
type bot interfaces{
	getGreeting() string
	// anyFunction() anyReturnType
}
```
The declaration of interfaces name bot have meaning:
Any types in the program that have function called `getGreeting` is member of bot interfaces.
It is mean... that if we have another function that receive our bot interfaces as it data type,
then we could sent all of bot interfaces member into that function.